Decode chipset and get the correct chipset string
This commit introduces a function to decode the chipset and convert it to
string before we are always get POWER8 as processor because it was hardcoded.
The decode chipset are used to set the package name and get some others
features. For now it's only used to get the package name.

Signed-off-by: Rogerio Alves <rcardoso@linux.vnet.ibm.com>